Feat 피드 content 수정 기능 구현

+++ src/main/java/project/closet/dto/response/FeedDto.java
@@ -12,8 +12,8 @@ public record FeedDto(
         WeatherSummaryDto weather,
         List<OotdDto> ootds,
         String content,
-        int likeCount,
-        int commentCount,
+        long likeCount,
+        long commentCount,
         boolean likedByMe
 ) {
